Simulates data from a graph

This is an internal simulation function for simulating data from a graph.

sim_X(B, n, sigma, alpha = rep(1, ncol(B)))

Arguments

B
a upper triangular matrix encoding a graph

n
number of observations from the graph

sigma
the common standard error

alpha
a vector of length equal to the number of parameters given by B. These alphas are multiplied to the common sigma.

Value

The sim_X function returns a matrix with data from the graph B with columns equal to the number of columns in B and rows equal to n.

Details

The sim_B function simulates a causal graph with regression coefficients. The graph must be encoded as an upper triangular matrix.

Examples

(B <- matrix(c(0,0,0,1,0,0,1,1,0), ncol = 3))

#>      [,1] [,2] [,3]
#> [1,]    0    1    1
#> [2,]    0    0    1
#> [3,]    0    0    0

sim_X(B, 10, 1)

#>              [,1]       [,2]       [,3]
#>  [1,] -0.55872582  0.6994304  1.0680653
#>  [2,] -1.39598601 -1.5372647 -2.4732923
#>  [3,] -1.36914864 -2.1249868 -1.8153382
#>  [4,]  0.82648974  4.0016095  5.7216303
#>  [5,] -0.49950853  0.7919735 -0.4439690
#>  [6,] -1.93412686 -1.5239191 -4.4655465
#>  [7,] -2.02669182 -0.9176423 -2.9052178
#>  [8,]  0.01352579  0.7840376  0.2429863
#>  [9,] -0.56260480 -1.4461271 -1.2742915
#> [10,]  0.69029351  1.0810565  0.1175527

sim_X(B, 10, 1, c(1, 5, 10))

#>             [,1]      [,2]       [,3]
#>  [1,] -0.4487055 -8.701124  -8.843050
#>  [2,]  1.5901592 -4.909776  -4.170367
#>  [3,] -0.3354367  5.972825  10.588898
#>  [4,]  1.6535075 -2.281024 -26.982157
#>  [5,]  0.6250739  8.016664  11.637409
#>  [6,] -0.8832099 -3.980372  -4.650104
#>  [7,]  0.5860456  6.946410   8.382256
#>  [8,] -0.3515753 -2.447098 -13.447637
#>  [9,] -1.5872343 -5.004272 -13.815413
#> [10,]  1.2179636  3.332936   5.148892